requirements.txt
Flask
Gunicorn
python-dotenv
bcrypt
firebase-admin